// apps/web/generated/api/models/projectRevisionRequestResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectRevisionRequestResponse {
  revision_id: string;
  project_id: string;
  project_delivery_id: string | null;
  requested_by_user_id: string;
  requested_to_user_id: string | null;
  round_no: number;
  status: string;
  reason: string;
  resolved_by_user_id: string | null;
  requested_at: string;
  resolved_at: string | null;
}